requirements.txt
pyspark==3.2.1

kafka-python==2.0.2